backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.0
celery>=5.3.6
python-multipart>=0.0.6
websockets>=12.0
pyyaml>=6.0.1
litellm>=1.17.0
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
